+++ cjs/index.d.ts
@@ -1,14 +1,59 @@
-declare class Timer {
+type TimerCallback<T = void> = null | (<R = T>(timer: Timer) => R);
+declare class Timer<T = void> {
     private __start;
     private __end;
     private __diff;
     private __closed;
     private __started;
     private __stopped;
+    private __label;
+    private __startCb;
+    private __stopCb;
+    private __closeCb;
     /**
      * Creates a new instance of {@link Timer}.
+     * @param label An optional label for this timer. Must be unique to this timer.
      */
-    constructor();
+    constructor(label?: string);
+    /**
+     * Resets the starting time, ending time, and difference.
+     *
+     * @returns `this` object for chaining.
+     */
+    reset(): this;
+    /**
+     * Delete the callback for the stopping of the timer.
+     * @returns `this` object for chaining.
+     */
+    deleteStopCb(): this;
+    /**
+     * Delete the callback for the starting of the timer.
+     * @returns `this` object for chaining.
+     */
+    deleteStartCb(): this;
+    /**
+     * Delete the callback for the closure of the timer.
+     * @returns `this` object for chaining.
+     */
+    deleteCloseCb(): this;
+    /**
+     * Register the callback for the stopping of the timer.
+     *
+     * @param cb A callback function, which will be invoked after the timer is stopped.
+     */
+    registerStopCb(cb: TimerCallback<T>): this;
+    /**
+     * Register the callback for the closure of the timer.
+     *
+     * @param cb A callback function, which will be invoked after the timer is closed.
+     */
+    registerCloseCb(cb: TimerCallback<T>): void;
+    /**
+     * Register the callback for the closure of the timer.
+     *
+     * @param cb A callback function, which will be invoked after the timer is started.
+     */
+    registerStartCb(cb: TimerCallback<T>): void;
     /**
      * Starts the timer.
      * @returns `this` object for chaining.
@@ -31,6 +76,11 @@ declare class Timer {
      */
     getDifference(): number;
     /**
+     * Reads this timer's label.
+     * @returns This timer's label.
+     */
+    getLabel(): string | null;
+    /**
      * Closes the timer so it can no longer be used.
      * @returns `this` object for chaining.
      */
@@ -51,6 +101,12 @@ declare class Timer {
      */
     isStopped(): boolean;
 }
+/**
+ * Creates a new instance of {@link Timer}.
+ * @param label An optional label. See {@link Timer}.
+ * @returns A new instance of `Timer`.
+ */
+declare function createTimer<T = void>(label?: string): Timer<T>;
 export default Timer;
-export { Timer };
+export { Timer, createTimer };
 export * as promises from './promises';